LayoutTests/imported/w3c/web-platform-tests/html/anonymous-iframe/hasStorageAccess.tentative.https.window-expected.txt


FAIL Same-origin credentialless iframe can't request storage access assert_equals: expected "false" but got "true"
PASS Cross-origin credentialless iframe can't request storage access
